main-component/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0
redis>=4.0.0
pymysql>=1.0.2
pytest>=6.2.5
locust>=2.8.3
black>=21.7b0
flake8>=3.9.0
isort>=5.9.3
SQLAlchemy==2.0.38